replace alert by alert dialog component and replace border by ring in…

+++ src/Components/Common/DeleteModal.tsx
@@ -1,10 +1,15 @@
 import {
-  Dialog,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogMedia,
+  AlertDialogTitle,
+} from "@/components/ui/alert-dialog";
+import { Trash2Icon } from "lucide-react";
 
 import {
   Fragment,
@@ -13,7 +18,6 @@ import {
   type ReactNode,
   type SetStateAction,
 } from "react";
-import RemoveIconSVG from "../Icons/RemoveIconSVG";
 
 const DeleteModal = ({
   modalTitle,
@@ -32,30 +36,21 @@ const DeleteModal = ({
       {/* Render the trigger with click handler */}
       {trigger(setOpen)}
 
-      <Dialog modal={true} open={open} onOpenChange={setOpen}>
-        <DialogContent
-          className="bg-white border-0 px-0"
-          showCloseButton={false}
-        >
-          <DialogHeader className="px-4 mb-4 pb-4 border-b border-gray-100 flex-row items-center justify-between">
-            <DialogTitle>{modalTitle}</DialogTitle>
+      <AlertDialog open={open} onOpenChange={setOpen}>
+        <AlertDialogContent className="bg-white border-0 px-0">
+          <AlertDialogHeader className="flex flex-col sm:group-data-[size=default]/alert-dialog-content:place-items-center">
+            <AlertDialogMedia className="bg-red-200 w-10 h-10">
+              <Trash2Icon className="text-red w-7 h-7" />
+            </AlertDialogMedia>
 
-            <button
-              type="button"
-              className="group cursor-pointer text-gray-500 transition-colors duration-350 hover:text-gray-700"
-              aria-label="Close Modal"
-              onClick={() => setOpen(false)}
-            >
-              <RemoveIconSVG
-                classNames="w-4 h-4"
-                pathClassNames="fill-gray-500 group-hover:fill-gray-700 transition-colors duration-350"
-              />
-            </button>
-          </DialogHeader>
-          <span className="px-4 mb-4">{modalDescription}</span>
-          <DialogFooter className="border-t border-gray-100 pt-4 px-4">
-            <button
-              type="button"
+            <AlertDialogTitle className="text-center">
+              {modalTitle}
+            </AlertDialogTitle>
+            <AlertDialogDescription>{modalDescription}</AlertDialogDescription>
+          </AlertDialogHeader>
+
+          <AlertDialogFooter className="border-t border-gray-100 pt-4 px-4">
+            <AlertDialogAction
               className="cursor-pointer bg-red text-white rounded-md px-4 py-2 transition-colors duration-350 hover:bg-rose-500"
               onClick={() => {
                 onDelete();
@@ -63,17 +58,17 @@ const DeleteModal = ({
               }}
             >
               Delete
-            </button>
-            <button
-              type="button"
+            </AlertDialogAction>
+
+            <AlertDialogCancel
               className="cursor-pointer bg-gray-500 text-white rounded-md px-4 py-2 transition-colors duration-350 hover:bg-gray-600"
               onClick={() => setOpen(false)}
             >
               Close
-            </button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+            </AlertDialogCancel>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
     </Fragment>
   );
 };
